+++ src/types/config.ts
@@ -1,0 +1,9 @@
+/**
+ * Color configuration types
+ */
+export type ColorName = 'yellow' | 'red' | 'green' | 'blue' | 'gray' | 'cyan' | 'magenta';
+
+export type SpecialValueConfig = {
+  color: ColorName;
+  bold: boolean;
+};
