+++ src/types/index.ts
@@ -1,1 +1,3 @@
 export * from './environment.js';
+export * from './config.js';
+export * from './table.js';
